improved shiftmem, improved mamagemem source, vma in progress.

+++ include/sys/mem.h
@@ -27,5 +27,5 @@
 #include <sys/defs.h>
 
 void setmem(void *start, void *end, char value);
-void shiftmem(void *start, void *end, uint32_t offset);
+void shiftmem(void *start, void *end, long long offset);
 void copymem(void *dest, void *src, uint32_t length);
